game.ts-and-app.ts-typed
