+++ index.html
@@ -60,7 +60,7 @@
       </nav>
     </div>
 
-    <script src="src/app.js" type="module"></script>
+    <script src="src/app.ts" type="module"></script>
 
   </body>
 </html>
