+++ package.json
@@ -5,7 +5,7 @@
   "name": "tamagotchi",
   "description": "<h1 align=\"center\">   Tamagotchi 👾 </h1>",
   "version": "1.0.0",
-  "main": "src/app.js",
+  "main": "src/app.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
